Move HTTPFileRetriever into fetch_data module
Also add a test.
